TRP-7: refactor to hooks
